Move leaderboards classes to gameplay namespace for now
