+++ osu.Game/Extensions/ModelExtensions.cs
@@ -8,7 +8,7 @@ using osu.Game.IO;
 using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Rulesets;
 using osu.Game.Scoring;
-using osu.Game.Screens.Select.Leaderboards;
+using osu.Game.Screens.Play.Leaderboards;
 using osu.Game.Users;
 
 namespace osu.Game.Extensions
