+++ osu.Game/Online/Leaderboards/LeaderboardManager.cs
@@ -18,7 +18,7 @@ using osu.Game.Online.API.Requests;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Scoring;
-using osu.Game.Screens.Select.Leaderboards;
+using osu.Game.Screens.Play.Leaderboards;
 using Realms;
 
 namespace osu.Game.Online.Leaderboards
