+++ osu.Game/Screens/OnlinePlay/Multiplayer/MultiplayerPlayer.cs
@@ -17,8 +17,8 @@ using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Rooms;
 using osu.Game.Scoring;
 using osu.Game.Screens.Play;
+using osu.Game.Screens.Play.Leaderboards;
 using osu.Game.Screens.Ranking;
-using osu.Game.Screens.Select.Leaderboards;
 using osu.Game.Users;
 using osuTK;
