+++ osu.Game/Screens/Ranking/SoloResultsScreen.cs
@@ -12,7 +12,7 @@ using osu.Game.Extensions;
 using osu.Game.Online.API;
 using osu.Game.Online.Leaderboards;
 using osu.Game.Scoring;
-using osu.Game.Screens.Select.Leaderboards;
+using osu.Game.Screens.Play.Leaderboards;
 
 namespace osu.Game.Screens.Ranking
 {
